Adjustments and improvements

+++ src/routers/rpc/math_rpc.rs
@@ -2,36 +2,36 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct ParamsForAdd {
-    pub a: i16,
-    pub b: i16,
+    pub a: f32,
+    pub b: f32,
 }
-pub fn add(params: ParamsForAdd) -> i16 {
+pub fn add(params: ParamsForAdd) -> f32 {
     params.a + params.b
 }
 
 #[derive(Deserialize)]
 pub struct ParamsForSubtract {
-    pub a: i16,
-    pub b: i16,
+    pub a: f32,
+    pub b: f32,
 }
-pub fn subtract(params: ParamsForSubtract) -> i16 {
+pub fn subtract(params: ParamsForSubtract) -> f32 {
     params.a - params.b
 }
 
 #[derive(Deserialize)]
 pub struct ParamsForMultiply {
-    pub a: i16,
-    pub b: i16,
+    pub a: f32,
+    pub b: f32,
 }
-pub fn multiply(params: ParamsForMultiply) -> i16 {
+pub fn multiply(params: ParamsForMultiply) -> f32 {
     params.a * params.b
 }
 
 #[derive(Deserialize)]
 pub struct ParamsForDivide {
-    pub a: i16,
-    pub b: i16,
+    pub a: f32,
+    pub b: f32,
 }
-pub fn divide(params: ParamsForDivide) -> i16 {
+pub fn divide(params: ParamsForDivide) -> f32 {
     params.a / params.b
 }
